增加优惠券及积分功能,导出最新sql

+++ yshop-system/src/main/java/co/yixiang/modules/shop/rest/YxStoreProductController.java
@@ -46,7 +46,7 @@ public class YxStoreProductController {
     @PostMapping(value = "/yxStoreProduct")
     @PreAuthorize("hasAnyRole('ADMIN','YXSTOREPRODUCT_ALL','YXSTOREPRODUCT_CREATE')")
     public ResponseEntity create(@Validated @RequestBody YxStoreProduct resources){
-        //if(ObjectUtil.isNotNull(resources)) throw new BadRequestException("演示环境禁止操作");
+        if(ObjectUtil.isNotNull(resources)) throw new BadRequestException("演示环境禁止操作");
         resources.setAddTime(OrderUtil.getSecondTimestampTwo());
         if(ObjectUtil.isEmpty(resources.getGiveIntegral())) resources.setGiveIntegral(BigDecimal.ZERO);
         if(ObjectUtil.isEmpty(resources.getCost())) resources.setCost(BigDecimal.ZERO);
@@ -71,7 +71,7 @@ public class YxStoreProductController {
     @DeleteMapping(value = "/yxStoreProduct/{id}")
     @PreAuthorize("hasAnyRole('ADMIN','YXSTOREPRODUCT_ALL','YXSTOREPRODUCT_DELETE')")
     public ResponseEntity delete(@PathVariable Integer id){
-        //if(id > 0) throw new BadRequestException("演示环境禁止操作");
+        if(id > 0) throw new BadRequestException("演示环境禁止操作");
         yxStoreProductService.delete(id);
         return new ResponseEntity(HttpStatus.OK);
     }
